fix payment request and response

+++ src/interfaces/payments/order.hpp
@@ -5,6 +5,6 @@ namespace gaboot
 	struct order
 	{
 		std::string order_id;
-		uint64_t gross_amount;
+		double gross_amount;
 	};
 }
